+++ packages/editor/editor-core/src/create-editor/ReactEditorView.tsx
@@ -21,8 +21,19 @@ import {
   AnalyticsDispatch,
   AnalyticsEventPayload,
   DispatchAnalyticsEvent,
+  ACTION,
+  ACTION_SUBJECT,
+  EVENT_TYPE,
+  FULL_WIDTH_MODE,
+  PLATFORMS,
+  AnalyticsEventPayloadWithChannel,
 } from '../plugins/analytics';
-import { EditorProps, EditorConfig, EditorPlugin } from '../types';
+import {
+  EditorProps,
+  EditorConfig,
+  EditorPlugin,
+  EditorAppearance,
+} from '../types';
 import { PortalProviderAPI } from '../ui/PortalProvider';
 import {
   pluginKey as editorDisabledPluginKey,
@@ -102,16 +113,14 @@ export default class ReactEditorView<T = {}> extends React.Component<
     if (allowAnalyticsGASV3) {
       this.activateAnalytics(createAnalyticsEvent);
     }
-
-    this.eventDispatcher.emit(analyticsEventKey, {
-      payload: {
-        action: 'started',
-        actionSubject: 'editor',
-        attributes: { platform: 'web' },
-        eventType: 'ui',
-      },
-    });
     initAnalytics(props.editorProps.analyticsHandler);
+
+    this.dispatchAnalyticsEvent({
+      action: ACTION.STARTED,
+      actionSubject: ACTION_SUBJECT.EDITOR,
+      attributes: { platform: PLATFORMS.WEB },
+      eventType: EVENT_TYPE.UI,
+    });
   }
 
   private broadcastDisabled = (disabled: boolean) => {
@@ -156,12 +165,32 @@ export default class ReactEditorView<T = {}> extends React.Component<
       }
     }
 
-    if (
-      nextProps.editorProps.appearance !== this.props.editorProps.appearance
-    ) {
+    const { appearance } = this.props.editorProps;
+    const { appearance: nextAppearance } = nextProps.editorProps;
+    if (nextAppearance !== appearance) {
       this.reconfigureState(nextProps);
+      if (nextAppearance === 'full-width' || appearance === 'full-width') {
+        this.dispatchAnalyticsEvent({
+          action: ACTION.CHANGED_FULL_WIDTH_MODE,
+          actionSubject: ACTION_SUBJECT.EDITOR,
+          eventType: EVENT_TYPE.TRACK,
+          attributes: {
+            previousMode: this.formatFullWidthAppearance(appearance),
+            newMode: this.formatFullWidthAppearance(nextAppearance),
+          },
+        });
+      }
     }
   }
+
+  formatFullWidthAppearance = (
+    appearance: EditorAppearance | undefined,
+  ): FULL_WIDTH_MODE => {
+    if (appearance === 'full-width') {
+      return FULL_WIDTH_MODE.FULL_WIDTH;
+    }
+    return FULL_WIDTH_MODE.DEFAULT;
+  };
 
   reconfigureState = (props: EditorViewProps) => {
     if (!this.view) {
@@ -353,15 +382,15 @@ export default class ReactEditorView<T = {}> extends React.Component<
             'atlaskit.fabric.editor.invalidtransaction',
             { documents: JSON.stringify(documents) }, // V2 events don't support object properties
           );
-          this.eventDispatcher.emit(analyticsEventKey, {
-            payload: {
-              action: 'dispatchedInvalidTransaction',
-              actionSubject: 'editor',
-              eventType: 'operational',
-              attributes: {
-                analyticsEventPayloads: transaction.getMeta(analyticsPluginKey),
-                documents,
-              },
+          this.dispatchAnalyticsEvent({
+            action: ACTION.DISPATCHED_INVALID_TRANSACTION,
+            actionSubject: ACTION_SUBJECT.EDITOR,
+            eventType: EVENT_TYPE.OPERATIONAL,
+            attributes: {
+              analyticsEventPayloads: transaction.getMeta(
+                analyticsPluginKey,
+              ) as AnalyticsEventPayloadWithChannel[],
+              documents,
             },
           });
         }