+++ packages/editor/editor-core/src/plugins/hyperlink/pm-plugins/input-rule.ts
@@ -50,6 +50,7 @@ export function createLinkInputRule(
           pos: start - (link.input!.length - link.lastIndex),
           appearance: 'inline',
           compareLinkText: true,
+          source: INPUT_METHOD.AUTO_DETECT,
         },
       ])(
         state.tr