+++ packages/media/media-ui/src/InlineCard/IconAndTitleLayout/index.tsx
@@ -1,5 +1,9 @@
 import * as React from 'react';
-import { IconTitleWrapper, IconWrapper } from './styled';
+import {
+  IconTitleWrapper,
+  IconWrapper,
+  IconTitleHeadNoBreakWrapper,
+} from './styled';
 import { Icon } from '../Icon';
 
 export interface IconAndTitleLayoutProps {
@@ -7,6 +11,8 @@ export interface IconAndTitleLayoutProps {
   title: React.ReactNode;
   right?: React.ReactNode;
 }
+
+const CHAR_LENGTH_BREAK_AT = 7;
 
 export class IconAndTitleLayout extends React.Component<
   IconAndTitleLayoutProps
@@ -29,10 +35,32 @@ export class IconAndTitleLayout extends React.Component<
 
   render() {
     const { title } = this.props;
+
+    // Check our text is actually a title text and not an element (e.g. as in UnauthorizedView)
+    if (typeof title === 'string') {
+      const head = title.slice(0, CHAR_LENGTH_BREAK_AT);
+      const rest = title.slice(CHAR_LENGTH_BREAK_AT);
+
+      return (
+        <>
+          <IconTitleWrapper>
+            <IconTitleHeadNoBreakWrapper>
+              {this.renderIcon()}
+              {head}
+            </IconTitleHeadNoBreakWrapper>
+            {rest}
+          </IconTitleWrapper>
+        </>
+      );
+    }
+
     return (
       <>
-        {this.renderIcon()}
-        <IconTitleWrapper>{title}</IconTitleWrapper>
+        <IconTitleWrapper>
+          {this.renderIcon()}
+          {'\u00A0'}
+          {title}
+        </IconTitleWrapper>
       </>
     );
   }