+++ packages/core/textarea/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/textarea
 
+## 1.0.0
+- [major] [dd95622388](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/dd95622388):
+
+  - This major release indicates that this package is no longer under dev preview but is ready for use
+
 ## 0.4.6
 - [patch] [0a4ccaafae](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0a4ccaafae):