+++ packages/elements/task-decision/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/task-decision
 
+## 14.0.7
+- [patch] [a6fb248987](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/a6fb248987):
+
+  - ED-6639 Align lists styles between editor & renderer
+
 ## 14.0.6
 - [patch] [0a4ccaafae](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0a4ccaafae):